Additional Functionality
Now their is an overlay for a win/tie. Where the correct winner is announced or a tie is announced and the user must interact with an overlay button to restart the game. Last thing needed to complete the project is give the user options for custom names and a way to reset the game at any point during the game.